+++ portal/code/cx-portal/src/state/api/userAdministration/userAdministrationAPI.ts
@@ -1,9 +1,9 @@
 import { HttpClient } from 'utils/HttpClient'
-import { InviteData } from 'types/userManagement/UserManagementTypes'
+import { InviteData } from 'types/userAdministration/UserAdministrationTypes'
 
-// Instance of UserManagement API endpoint
-export class UserManagementApi extends HttpClient {
-  private static classInstance?: UserManagementApi
+// Instance of UserAdministration API endpoint
+export class UserAdministrationApi extends HttpClient {
+  private static classInstance?: UserAdministrationApi
 
   // TODO: Token needs to read from Redux store
   public constructor(token: string) {
@@ -15,7 +15,7 @@ export class UserManagementApi extends HttpClient {
   // To avoid create an instance everytime, pointed to Singleton of static value
   public static getInstance(token: string) {
     if (!this.classInstance) {
-      this.classInstance = new UserManagementApi(token)
+      this.classInstance = new UserAdministrationApi(token)
     }
 
     return this.classInstance
